# api/openapi.yaml
openapi: "3.1.0"
info:
  title: FamilyRoots API
  version: "1.0.0"
  description: |
    ## FamilyRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Family tree identifier
...
    mediaId:
      name: mediaId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Media item identifier
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Maximum number of items to return
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response

  # ──────────────────────────────────────────────────────────
  # COMMON HEADERS
  # ──────────────────────────────────────────────────────────
...
  schemas:
...
    MediaType:
      type: string
      enum: [PHOTO, DOCUMENT, AUDIO, VIDEO, OTHER]
...
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://familyroots.app/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
    Pagination:
      type: object
      required: [hasMore]
      properties:
        cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page (null if no more pages)
        hasMore:
          type: boolean
        totalCount:
          type: integer
          description: Total matching records (expensive; only returned when requested)

    # ── Auth / Tokens ───────────────────────────────────────
...
    MediaUploadRequest:
      type: object
      required: [filename, mimeType, fileSizeBytes]
      properties:
        filename:
          type: string
          maxLength: 255
        mimeType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/gif
            - image/webp
            - image/heic
            - application/pdf
            - audio/mpeg
            - audio/wav
            - video/mp4
            - video/quicktime
        fileSizeBytes:
          type: integer
          minimum: 1
          maximum: 524288000
          description: Maximum 500 MB
      description: Request a pre-signed upload URL. Upload the file directly to S3 using the returned URL.
    MediaUploadResponse:
      type: object
      required: [mediaId, uploadUrl, expiresAt]
      properties:
        mediaId:
          type: string
          format: uuid
          description: Media record created in PENDING status
        uploadUrl:
          type: string
          format: uri
          description: Pre-signed S3 PUT URL — valid for 15 minutes
        expiresAt:
          type: string
          format: date-time
    Media:
      type: object
      required: [id, mediaType, originalFilename, status, fileSizeBytes, mimeType, createdAt]
      properties:
        id:
          type: string
          format: uuid
        mediaType:
          $ref: "#/components/schemas/MediaType"
        originalFilename:
          type: string
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        variants:
          type: array
          items:
            $ref: "#/components/schemas/MediaVariant"
        fileSizeBytes:
          type: integer
        mimeType:
          type: string
        widthPx:
          type: integer
          nullable: true
        heightPx:
          type: integer
          nullable: true
        status:
          type: string
          enum: [PENDING, PROCESSING, READY, FAILED, DELETED]
        uploadedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MediaVariant:
      type: object
      required: [size, url, width, height]
      properties:
        size:
          type: string
          example: "800w"
        url:
          type: string
          format: uri
        width:
          type: integer
        height:
          type: integer
...
  responses:
    400BadRequest:
      description: Bad Request — malformed JSON or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403Forbidden:
      description: Forbidden — valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /trees/{treeId}/media:
    parameters:
      - $ref: "#/components/parameters/treeId"
    get:
      tags: [Media]
      operationId: listMedia
      summary: List media items in a tree
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
        - name: mediaType
          in: query
          schema:
            $ref: "#/components/schemas/MediaType"
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PROCESSING, READY, FAILED]
      responses:
        "200":
          description: Paginated media items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Media"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"

    post:
      tags: [Media]
      operationId: requestMediaUpload
      summary: Request a pre-signed upload URL for a media file
      description: >
        Two-step upload flow:
        1. POST here to receive a pre-signed S3 PUT URL and a mediaId.
        2. PUT the file binary directly to the S3 URL (client-to-S3, no API proxy).
        3. The media record transitions from PENDING → PROCESSING → READY automatically.
        Poll GET /trees/{treeId}/media/{mediaId} to check status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadRequest"
      responses:
        "201":
          description: Pre-signed URL issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResponse"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "413":
          description: File size exceeds plan limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /trees/{treeId}/media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/treeId"
      - $ref: "#/components/parameters/mediaId"
    get:
      tags: [Media]
      operationId: getMedia
      summary: Get a media item by ID
      responses:
        "200":
          description: Media item detail with signed URLs for original and variants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    patch:
      tags: [Media]
      operationId: updateMedia
      summary: Update media metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                originalFilename:
                  type: string
                  maxLength: 255
      responses:
        "200":
          description: Updated media item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    delete:
      tags: [Media]
      operationId: deleteMedia
      summary: Soft-delete a media item
      description: Marks the record as DELETED; S3 object is removed by a background job
      responses:
        "204":
          description: Media deleted
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
...